+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.6] / 2022-11-30 - 2022-10-31
+### Features
+- Add `TestEngineFilter` simple filter by method/test name
+- Add SampleTest Project
+- Local Test with `SampleTest` in Revit
+- Skipped Test Parameter
+- Test with `ExpectedResult` feature
+### Tests
+- Add `ricaun.NUnit.Tests`
+- Add `SampleTest.Tests`
+### Added
+- Add `ITest` in Build
+- Add `GetMethodTestNames` - Get Test Names
+- Add `IsValueExpectedResult` case expected result
+- Add `Skipped` feature
+- Add `IgnoreException` feature
+
 ## [1.0.5] / 2022-09-20
 ### Features
 - Remove await / async
@@ -42,6 +59,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Test with Console Output
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.0.6]: ../../compare/1.0.5...1.0.6
+[1.0.5]: ../../compare/1.0.4...1.0.5
+[1.0.4]: ../../compare/1.0.3...1.0.4
 [1.0.3]: ../../compare/1.0.2...1.0.3
 [1.0.2]: ../../compare/1.0.1...1.0.2
 [1.0.1]: ../../compare/1.0.0...1.0.1